+++ Example/RadioGroupDemo/ViewController.swift
@@ -17,11 +17,11 @@ class ViewController: UIViewController {
         super.viewDidLoad()
 
         radioGroup.titles = ["One", "Two", "Last"]
-        secondGroup.titles = [
-            "Lorem ipsum dolor sit amet",
-            "consectetur adipiscing elit",
-            "sed do eiusmod tempor incididunt",
-            "ut labore et dolore magna aliqua",
+        secondGroup.attributedTitles = [
+            NSAttributedString(string: "Lorem ipsum dolor sit amet"),
+            NSAttributedString(string: "consectetur adipiscing elit", attributes: [.strokeWidth: -3.0]),
+            NSAttributedString(string: "sed do eiusmod tempor incididunt", attributes: [.obliqueness: 0.3]),
+            NSAttributedString(string: "ut labore et dolore magna aliqua", attributes: [.underlineStyle: 2, .underlineColor: secondGroup.tintColor ?? .orange]),
         ]
 
         radioGroup.addTarget(self, action: #selector(didSelectOption(radioGroup:)), for: .valueChanged)
@@ -31,6 +31,6 @@ class ViewController: UIViewController {
     }
 
     @objc func didSelectOption(radioGroup: RadioGroup) {
-        print(radioGroup.titles[radioGroup.selectedIndex])
+        print(radioGroup.titles[radioGroup.selectedIndex] ?? "")
     }
 }