+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.comd/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.orgd/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.3.3] - 2019-09-12
 
 ### Fixed
 - calculate `intrinsicContentSize` to make the group work better with SwiftUI `fixedSize()`.